LayoutTests/http/tests/security/contentSecurityPolicy/1.1/securitypolicy-allowconnectionto-expected.txt

PASS connection is allowed when no policy exists.
PASS connection is not allowed when policy exists.
